clean up

+++ src/main/java/org/resala/core/volunteer/repository/VolunteerRegistrationRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface VolunteerRegistrationRepository extends CrudRepository<VolunteerRegistrationEntity, VolunteerRegistrationKey> {
 
-    long countByBranchIdAndEventId(long branchId, Long eventId);
+    int countByBranchIdAndEventId(long branchId, Long eventId);
 }
